backend/requirements.txt
fastapi
uvicorn
trafilatura
nltk
networkx
scikit-learn
pymongo
pydantic
requests
python-dotenv